code/requirements.txt
opencv_python
matplotlib
numpy
scipy
tensorflow_gpu==2.1.0
tqdm
PyQt5
scikit_learn

